+++ dev/mathbox/examples/math/color_cube.clj
@@ -3,7 +3,14 @@
  :no-cache true
  :visibility :hide-ns}
 (ns mathbox.examples.math.color-cube
-  (:require [mentat.clerk-utils.show :refer [show-sci]]))
+  (:require [mentat.clerk-utils.show :refer [show-sci]]
+            [nextjournal.clerk :as clerk]))
+
+^{::clerk/visibility {:code :hide :result :hide}}
+(clerk/eval-cljs
+ ;; These aliases only apply inside this namespace.
+ '(require '[mathbox.core :as mathbox])
+ '(require '[mathbox.primitives :as mb]))
 
 ;; ### Color Cube